requirements.txt
google-generativeai
ollama
markitdown
youtube-transcript-api
python-dotenv
